requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-allauth==0.45.0
django-bootstrap-form==3.4
django-bootstrap-forms==0.1
django-bootstrap4==3.0.1
django-forms-bootstrap==3.1.0
future==0.18.2
html5lib==1.1
idna==3.11
oauthlib==3.3.1
Pillow==8.4.0
pycparser==2.23
PyJWT==2.10.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
reportlab==3.6.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
typing-extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.5
